pyproject.toml
[project]
name = "ainc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.20.0",
    "litellm>=1.72.6",
    "numpy>=2.3.0",
    "python-dotenv>=1.1.0",
]
